refactor: Move background service and DI extensions to Capsule.Core
