+++ sources/Capsule.Test/AutomatedTests/UnitTests/DisposeTest.cs
@@ -1,5 +1,5 @@
 ﻿using Capsule.Attribution;
-using Capsule.Extensions.DependencyInjection;
+using Capsule.GenericHosting;
 
 using Microsoft.Extensions.Logging;
